Update RTC.cpp
am_util_time_computeDayofWeek function needs a month number comprised between 1 and 12.
There's an error on line 62 : month argument is already in that gap. It must not be added 1.

It's not the same case at line 77 : mthToIndex(&__DATE__[0]) returns a month number comprised between 0 and 11, this the reason why that argument needs to be added 1.

Those two lines are similar but 2nd argument passed to am_util_time_computeDayofWeek function is not the same although it looks the same.